LinBPQ 6.0.20
Single commit of the released version as available on the .zip at
https://www.cantab.net/users/john.wiseman/Downloads/ .

+++ BPQINP3.c
@@ -828,7 +828,7 @@ int BuildRIF(UCHAR * RIF, UCHAR * Call, UCHAR * Alias, int Hops, int RTT)
 		if (ptr)
 			*ptr = 0;
 
-		AliasLen = strlen(AliasCopy);
+		AliasLen = (int)strlen(AliasCopy);
 
 		RIF[10] = AliasLen+2;
 		RIF[11] = 0;
